Updates dzToon.lua.
Enhance DEBUG logging and removed  if isTimer clause.

Now the script will performance the actions if a user changed program state or switch and immediately query Toon for the rest of the devices.
E.g. a change in the setpoint will immediately change the program state to temporary and other descendants.